API functions for weather data are in progress
