+++ source/backend/api/__init__.py
@@ -1,16 +1,11 @@
-import datetime
-
-from config import settings
-from backend.service.utils.get_db_session import get_db_session
-from backend.service.db.repositories.user_repository import UserRepository
-from backend.service.db.repositories.geographical_unit_repository import GeographicalUnitRepository
-from backend.service.db.repositories.load_data_repository import LoadDataRepository
-from backend.service.db.repositories.weather_repository import WeatherRepository
-from backend.service.db.entities.geographical_unit_entities import GeographicalUnitEntity
 from backend.service.db.models.enums import GeographicalUnitCode, GeographicalUnitType, RegulatorType
-
-from backend.service.data_clients.entsoe import ENTSOEClient
-from backend.service.data_clients.asos import ASOSClient
+from backend.service.db.repositories.geographical_unit_repository import GeographicalUnitRepository
+from backend.service.db.repositories.user_repository import UserRepository
+from backend.service.utils.get_db_session import get_db_session
+from config import settings
+from .geo_unit import add_new_geographical_unit, update_geographical_unit, get_geographical_unit
+from .load_data import add_new_load_data, fetch_and_add_new_load_data, sync_load_data
+from .weather import fetch_and_add_weather_stations
 
 
 def add_new_user(user: dict):
@@ -22,188 +17,6 @@ def add_new_user(user: dict):
         print(e)
     else:
         session.close()
-
-
-def add_new_geographical_unit(geographical_unit: dict):
-    """
-    geographical_unit = {
-        'code': GeographicalUnitCode.TURKIYE,
-        'name': 'Turkiye',
-        'type': GeographicalUnitType.COUNTRY,
-        'regulator': RegulatorType.EPIAS,
-        'is_active': True,
-        'created_by_id': 1,
-        'updated_by_id': 1,
-        'created_at': datetime.datetime.now(datetime.timezone.utc),
-        'updated_at': datetime.datetime.now(datetime.timezone.utc),
-    }
-    """
-
-    session = get_db_session(database_url=settings.DATABASE_URL)
-    geographical_unit_repository = GeographicalUnitRepository(session=session)
-    try:
-        geographical_unit_repository.add_new_geographical_unit(geographical_unit=geographical_unit)
-    except RuntimeError as e:
-        print(e)
-    else:
-        session.close()
-
-
-def update_geographical_unit(
-        code: GeographicalUnitCode,
-        regulator: RegulatorType,
-        last_valid_data_ending: datetime.datetime
-):
-    session = get_db_session(database_url=settings.DATABASE_URL)
-    geographical_unit_repository = GeographicalUnitRepository(session=session)
-    try:
-        geographical_unit_repository.update_geographical_unit(code=code, regulator=regulator,
-                                                              last_valid_data_ending=last_valid_data_ending)
-    except RuntimeError as e:
-        print(e)
-    else:
-        session.close()
-
-
-def fetch_and_add_weather_stations(code: GeographicalUnitCode, regulator: RegulatorType):
-    session = get_db_session(database_url=settings.DATABASE_URL)
-    geographical_unit_repository = GeographicalUnitRepository(session=session)
-    try:
-        geo_unit = geographical_unit_repository.get_geographical_unit_from_code(code=code, regulator=regulator)
-    except RuntimeError as e:
-        print(e)
-        raise RuntimeError(e)
-    else:
-        session.close()
-
-    asos_client = ASOSClient()
-    stations = asos_client.get_stations_for_network(country_code=code.value)
-
-    for station in stations:
-        station['geographical_unit_id'] = geo_unit.id
-        station['created_at'] = datetime.datetime.now(datetime.timezone.utc)
-        station['updated_at'] = datetime.datetime.now(datetime.timezone.utc)
-        station['created_by_id'] = 1
-        station['updated_by_id'] = 1
-
-    session = get_db_session(database_url=settings.DATABASE_URL)
-    weather_repository = WeatherRepository(session=session)
-    try:
-        weather_repository.add_new_weather_stations(weather_stations=stations)
-    except RuntimeError as e:
-        print(e)
-        raise RuntimeError(e)
-    else:
-        session.close()
-
-
-def add_new_load_data(
-        load_data: list[dict],
-        geographical_unit_code: GeographicalUnitCode,
-        regulator: RegulatorType,
-):
-    """
-    load_data: dict or list of dict
-    """
-
-    session = get_db_session(database_url=settings.DATABASE_URL)
-    load_data_repository = LoadDataRepository(session=session)
-    try:
-        load_data_repository.add_new_load_data(load_data=load_data)
-    except RuntimeError as e:
-        print(e)
-    else:
-        session.close()
-
-    last_valid_data_ending = load_data[-1]['end_datetime']  # data points must be ordered
-    update_geographical_unit(code=geographical_unit_code,
-                             regulator=regulator,
-                             last_valid_data_ending=last_valid_data_ending)
-
-
-def get_geographical_unit(entity_code: GeographicalUnitCode, regulator: RegulatorType) -> GeographicalUnitEntity:
-    session = get_db_session(database_url=settings.DATABASE_URL)
-    geographical_unit_repository = GeographicalUnitRepository(session=session)
-    geographical_unit = None
-
-    try:
-        geographical_unit = geographical_unit_repository.get_geographical_unit_from_code(code=entity_code,
-                                                                                         regulator=regulator)
-    except RuntimeError as e:
-        print(e)
-    else:
-        session.close()
-
-    return geographical_unit
-
-
-def fetch_and_add_new_load_data(
-        entity_code: GeographicalUnitCode,
-        regulator: RegulatorType,
-        start_datetime: datetime.datetime,
-        end_datetime: datetime.datetime
-):
-    geographical_unit = get_geographical_unit(entity_code=entity_code, regulator=regulator)
-
-    match regulator:
-        case RegulatorType.ENTSOE:
-            data_client = ENTSOEClient(token=settings.ENTSOE_TOKEN)
-        case _:
-            raise RuntimeError('Only ENTSOE Data Client is implemented!')
-
-    data_list = data_client.get_load_data(entity_code=entity_code.value,
-                                          start_datetime=start_datetime,
-                                          end_datetime=end_datetime)
-
-    if len(data_list) > 0:
-        # Updates data_list in-place
-        _ = [
-            x.update(
-                {
-                    'created_by_id': 1,
-                    'updated_by_id': 1,
-                    'created_at': datetime.datetime.now(datetime.timezone.utc),
-                    'updated_at': datetime.datetime.now(datetime.timezone.utc),
-                    'geographical_unit_id': geographical_unit.id,
-                }
-            ) for x in data_list
-        ]
-        add_new_load_data(load_data=data_list, geographical_unit_code=entity_code, regulator=regulator)
-
-
-def sync_load_data(entity_code: GeographicalUnitCode, regulator: RegulatorType):
-    MAX_QUERY_DAYS = 180
-    ABSOLUTE_BEGINNING = datetime.datetime(year=2015, month=1, day=1, hour=0, minute=0, tzinfo=datetime.timezone.utc)
-
-    geographical_unit = get_geographical_unit(entity_code=entity_code, regulator=regulator)
-
-    if geographical_unit.last_valid_data_ending is None:
-        geographical_unit.last_valid_data_ending = ABSOLUTE_BEGINNING
-
-    current_datetime = datetime.datetime.now(datetime.timezone.utc).replace(minute=0, second=0, microsecond=0)
-
-    while current_datetime - geographical_unit.last_valid_data_ending > datetime.timedelta(days=MAX_QUERY_DAYS):
-        start_datetime = geographical_unit.last_valid_data_ending
-        end_datetime = start_datetime + datetime.timedelta(days=MAX_QUERY_DAYS)
-        prev_last_valid_data_ending = geographical_unit.last_valid_data_ending
-
-        fetch_and_add_new_load_data(entity_code=entity_code,
-                                    regulator=regulator,
-                                    start_datetime=start_datetime,
-                                    end_datetime=end_datetime)
-        geographical_unit = get_geographical_unit(entity_code=entity_code, regulator=regulator)
-
-        if ((geographical_unit.last_valid_data_ending is None) or
-                (geographical_unit.last_valid_data_ending == prev_last_valid_data_ending)):
-            geographical_unit.last_valid_data_ending = end_datetime
-
-    else:  # of WHILE
-        start_datetime = geographical_unit.last_valid_data_ending
-        end_datetime = current_datetime
-        fetch_and_add_new_load_data(entity_code=entity_code,
-                                    regulator=regulator,
-                                    start_datetime=start_datetime,
-                                    end_datetime=end_datetime)
 
 
 def sync_all_data():
@@ -220,9 +33,9 @@ def sync_all_data():
     else:
         session.close()
 
-    for geo_unit in geographical_units:
-        print(f'Syncing: {geo_unit.name} - {geo_unit.code.value}')
-        sync_load_data(entity_code=geo_unit.code, regulator=RegulatorType.ENTSOE)
+    for g_unit in geographical_units:
+        print(f'Syncing: {g_unit.name} - {g_unit.code.value}')
+        sync_load_data(entity_code=g_unit.code, regulator=RegulatorType.ENTSOE)
 
 
 ################
