+++ source/backend/service/db/entities/weather_entities.py
@@ -19,3 +19,21 @@ class WeatherStationEntity:
     created_at: datetime.datetime
     updated_by_id: int
     updated_at: datetime.datetime
+
+
+@dataclass
+class WeatherDataEntity:
+    id: int
+    station_id: int
+    measured_at: datetime.datetime
+    temperature_celsius: float
+    dewpoint_celsius: float
+    relative_humidity: float
+    wind_direction: float
+    wind_speed: float
+    pressure_altimeter: float
+    feel_celsius: float
+    created_by_id: int
+    updated_by_id: int
+    created_at: datetime.datetime
+    updated_at: datetime.datetime
